refactor(cli): split row filtering out of fmt into filter module
fmt.rs mixed presentation (render/table/json) with query filtering (Where,
Op, apply_where, stats_exclude). Move the filter half to filter.rs verbatim;
pure relocation, no behavior change.
